add bundler plugin as rubygems plugin is not loaded on `bundle install`
`bundle install` does not run rubygems install hooks when it does *path*
installs. This is because bundler puts dependencies in some loadable path
but does not install them as gem on *path* install.

https://github.com/rubygems/bundler/issues/5429

To enable automatic hooks configuration in both `gem install` and
`bundle install` cases, this package needs to be installed as a bundler
plugin.